code

+++ app/Http/Controllers/API/APILichChieuController.php
@@ -3,29 +3,35 @@
 namespace App\Http\Controllers\API;
 
 use App\Http\Controllers\Controller;
+use App\Models\GheChieu;
 use App\Models\LichChieu;
 use App\Models\Phim;
 use App\Models\PhongChieu;
+use App\Models\VeXemPhim;
 use Carbon\Carbon;
+use Exception;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Str;
 
 class APILichChieuController extends Controller
 {
     public function data(Request $request)
     {
         $data       =   LichChieu::join('phims', 'phims.id', 'lich_chieus.id_phim')
-                                 ->join('phong_chieus', 'lich_chieus.id_phong', 'phong_chieus.id')
-                                 ->select('lich_chieus.*', 'phims.ten_phim', 'phong_chieus.ten_phong')
-                                 ->get();
+            ->join('phong_chieus', 'lich_chieus.id_phong', 'phong_chieus.id')
+            ->select('lich_chieus.*', 'phims.ten_phim', 'phong_chieus.ten_phong')
+            ->get();
 
         $today      =   Carbon::today();
 
         $ds_phim    =   Phim::where('hien_thi', 1)
-                            ->where('ket_thuc', '>', $today)
-                            ->get();
+            ->where('ket_thuc', '>', $today)
+            ->get();
 
         $ds_phong   =   PhongChieu::where('tinh_trang', 1)
-                                  ->get();
+            ->get();
 
         return response()->json([
             'data'      =>  $data,
@@ -36,61 +42,102 @@ class APILichChieuController extends Controller
 
     public function store(Request $request)
     {
-        $data   = $request->all();
+        DB::beginTransaction();
+        try {
+            $data   = $request->all();
 
-        LichChieu::create($data);
+            LichChieu::create($data);
 
-        return response()->json([
-            'status'    => 1,
-            'message'   => 'Đã thêm mới lịch chiếu thành công!',
-        ]);
+            DB::commit();
+            return response()->json([
+                'status'    => 1,
+                'message'   => 'Đã thêm mới lịch chiếu thành công!',
+            ]);
+        } catch (Exception $e) {
+            Log::error("Ê, nó có lỗi đó tề: " . $e);
+            DB::rollBack();
+        }
     }
 
     public function status(Request $request)
     {
-        $lich_chieu     = LichChieu::find($request->id);
+        DB::beginTransaction();
+        try {
+            $lich_chieu     = LichChieu::find($request->id);
+            if ($lich_chieu) {
+                // Nếu lịch chiếu đang chuyển từ hoạt động  => tạm tắt
+                if ($lich_chieu->trang_thai == 1) {
+                    // Kiểm tra xem đã có vé nào bán hay chưa?
+                    $check  = VeXemPhim::where('id_lich_chieu', $request->id)
+                                       ->where('tinh_trang', \App\Models\VeXemPhim::VE_DA_BAN)
+                                       ->first();
+                    if($check) {
+                        return response()->json([
+                            'status'    => 0,
+                            'message'   => 'Lịch chiếu này đã bán vé cho khách rồi!',
+                        ]);
+                    }
+                    // Phải hủy toàn bộ vé đã tạo ra
+                    VeXemPhim::where('id_lich_chieu', $request->id)->delete();
 
-        if($lich_chieu) {
-            // Nếu lịch chiếu đang chuyển từ hoạt động  => tạm tắt
-            if($lich_chieu->trang_thai == 1) {
-                $lich_chieu->trang_thai = 0;
-                $lich_chieu->save();
+                    $lich_chieu->trang_thai = 0;
+                    $lich_chieu->save();
 
-                return response()->json([
-                    'status'    => 1,
-                    'message'   => 'Đã hủy lịch chiếu phim!',
-                ]);
-            } else {
-                // B1: ta cần kiểm ta xem tại phòng này và thời gian này có chiếu phim khác chưa
-                // B2: nếu là chưa
-                $check  = random_int(0, 1);
+                    DB::commit();
 
-                if($check == false) {
+                    return response()->json([
+                        'status'    => 1,
+                        'message'   => 'Đã hủy lịch chiếu phim!',
+                    ]);
+                } else {
+                    // a. Kiểm tra lịch chiếu có trùng không?   => Quên đi và ta chưa code
+
+                    // b. Kiểm tra xem phòng chiếu này đã có ghế hay chưa?
+                    $gheChieu   =   GheChieu::where('id_phong_chieu', $request->id_phong)->get();
+
+                    if(count($gheChieu) == 0) {
+                        return response()->json([
+                            'status'    => 0,
+                            'message'   => 'Phòng chiếu này không có ghế để bán!',
+                        ]);
+                    }
+                    // c. Tạo ra danh sách ghế để bán. Ví dụ: Phòng 1 có 30 ghế thì ta sẽ tạo ra 30 ghế có thể bán cho lịch chiếu này. Nhưng ở ghế chiếu có trạng thái (0/1)	=> Chúng ta vẫn tạo đủ 30 ghế nhưng chúng ta chỉ bán những ghế có trạng thái = 1
+                    foreach($gheChieu as $key => $value) {
+                        VeXemPhim::create([
+                            'id_lich_chieu'     =>  $request->id,   // Lịch mà ta đang đổi trạng thái
+                            'so_ghe'            =>  $value->ten_ghe,
+                            'ma_ve'             =>  Str::uuid(),    // sinh ra 1 đoạn mã random 36 ký tự không trùng
+                            'gia_ve'            =>  $value->gia_mac_dinh,
+                            'tinh_trang'        =>  $value->tinh_trang == 0 ? \App\Models\VeXemPhim::VE_KHONG_THE_BAN : \App\Models\VeXemPhim::VE_CO_THE_BAN,
+                        ]);
+                    }
+
                     $lich_chieu->trang_thai = 1;
                     $lich_chieu->save();
+
+                    DB::commit();
 
                     return response()->json([
                         'status'    => 1,
                         'message'   => 'Đã kích hoạt lịch chiếu phim!',
                     ]);
-                } else {
-                    return response()->json([
-                        'status'    => 0,
-                        'message'   => 'Giờ này đã có lịch chiếu phim khác!',
-                    ]);
                 }
+            } else {
+                return response()->json([
+                    'status'    => 0,
+                    'message'   => 'Lịch chiếu không tồn tại!',
+                ]);
             }
-        } else {
-            return response()->json([
-                'status'    => 0,
-                'message'   => 'Lịch chiếu không tồn tại!',
-            ]);
+        } catch (Exception $e) {
+            Log::error($e);
+            DB::rollBack();
         }
     }
 
-    public function update(Request $request){
+    public function update(Request $request)
+    {
         $lichChieu   = LichChieu::find($request->id);
-        if($lichChieu) {
+        if ($lichChieu) {
             $data   = $request->all();
             $lichChieu->update($data);
 
@@ -105,21 +152,45 @@ class APILichChieuController extends Controller
             ]);
         }
     }
-    public function destroy(Request $request){
-        $lichChieu     =   LichChieu::find($request->id);
 
-        if($lichChieu) {
-            $lichChieu->delete();
+    public function destroy(Request $request)
+    {
+        DB::beginTransaction();
+        try {
+            $lichChieu     =   LichChieu::find($request->id);
 
-            return response()->json([
-                'status'    => 1,
-                'message'   => 'Đã xóa lịch chiếu thành công!',
-            ]);
-        } else {
-            return response()->json([
-                'status'    => 0,
-                'message'   => 'lịch chiếu không tồn tại!',
-            ]);
+            if ($lichChieu) {
+                // Kiểm tra xem đã có vé nào bán hay chưa?
+                $check  = VeXemPhim::where('id_lich_chieu', $request->id)
+                                   ->where('tinh_trang', \App\Models\VeXemPhim::VE_DA_BAN)
+                                   ->first();
+                if($check) {
+                    return response()->json([
+                        'status'    => 0,
+                        'message'   => 'Lịch chiếu này đã bán vé cho khách rồi!',
+                    ]);
+                }
+                // Phải hủy toàn bộ vé đã tạo ra
+                VeXemPhim::where('id_lich_chieu', $request->id)->delete();
+
+                $lichChieu->delete();
+
+                DB::commit();
+
+                return response()->json([
+                    'status'    => 1,
+                    'message'   => 'Đã xóa lịch chiếu thành công!',
+                ]);
+            } else {
+                return response()->json([
+                    'status'    => 0,
+                    'message'   => 'lịch chiếu không tồn tại!',
+                ]);
+            }
+
+        } catch(Exception $e) {
+            Log::error($e);
+            DB::rollBack();
         }
     }
 }
